Upgrade elvis, relint, extend CI matrix
I had to upgrade elvis in order to make it work on OTP 27.

It introduced some new rules that I've implemented.
A note about "state_record_and_type" is that it seems to be deprecated in the official ruleset so I took it out of the config since it was causing issues.
- export_used_types was complaining about state/0
- private_data_types was complaining that it wasn't opaque
- then state_record_and_type was complaining that it was missing
Seems to be related:
inaka/elvis_core#171